# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/attribute": {
      "post": {
        "description": "Chỉ dành cho tài khoản có quyền Admin. Dùng để định nghĩa các thuộc tính mở rộng cho phân quyền.",
        "operationId": "AttributeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttributeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ResponseAttributeDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 400,
                      "message": "Bad Request",
                      "requestId": "a0b97693-ec90-4797-a1f9-a76b379da91c",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 404,
                      "message": "Not Found",
                      "requestId": "3793ab9a-5673-4c20-8580-c3240e3f3342",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 409,
                      "message": "Conflict",
                      "requestId": "a3d8f67a-a6d6-4a98-ad6c-ffa15711e862",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 500,
                      "message": "Internal Server Error",
                      "requestId": "a9b02586-d72b-41b3-be78-1d1f29e18312",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo điều kiện quyền mới",
        "tags": [
          "Attribute (Quản lý thuộc tính/điều kiện)"
        ]
      },
      "get": {
        "description": "Hỗ trợ lọc theo tên, sắp xếp và phân trang.",
        "operationId": "AttributeController_findAllPaginated",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Tìm kiếm theo tên hoặc mô tả của điều kiện.",
            "schema": {
              "example": "department",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang hiện tại.",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bản ghi trên mỗi trang.",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Trường dữ liệu dùng để sắp xếp.",
            "schema": {
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Thứ tự sắp xếp.",
            "schema": {
              "default": "asc",
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ResponseAttributeDto"
                          }
                        },
                        "meta": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "number"
                            },
                            "page": {
                              "type": "number"
                            },
                            "limit": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 400,
                      "message": "Bad Request",
                      "requestId": "a0b97693-ec90-4797-a1f9-a76b379da91c",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 404,
                      "message": "Not Found",
                      "requestId": "3793ab9a-5673-4c20-8580-c3240e3f3342",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 409,
                      "message": "Conflict",
                      "requestId": "a3d8f67a-a6d6-4a98-ad6c-ffa15711e862",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 500,
                      "message": "Internal Server Error",
                      "requestId": "a9b02586-d72b-41b3-be78-1d1f29e18312",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách điều kiện quyền (phân trang)",
        "tags": [
          "Attribute (Quản lý thuộc tính/điều kiện)"
        ]
      }
    },
    "/attribute/{id}": {
      "get": {
        "operationId": "AttributeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID của điều kiện quyền (định dạng UUID)",
            "schema": {
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ResponseAttributeDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 400,
                      "message": "Bad Request",
                      "requestId": "a0b97693-ec90-4797-a1f9-a76b379da91c",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 404,
                      "message": "Not Found",
                      "requestId": "3793ab9a-5673-4c20-8580-c3240e3f3342",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 409,
                      "message": "Conflict",
                      "requestId": "a3d8f67a-a6d6-4a98-ad6c-ffa15711e862",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 500,
                      "message": "Internal Server Error",
                      "requestId": "a9b02586-d72b-41b3-be78-1d1f29e18312",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy chi tiết một điều kiện quyền",
        "tags": [
          "Attribute (Quản lý thuộc tính/điều kiện)"
        ]
      },
      "patch": {
        "operationId": "AttributeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID của điều kiện quyền (định dạng UUID)",
            "schema": {
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttributeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ResponseAttributeDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 400,
                      "message": "Bad Request",
                      "requestId": "a0b97693-ec90-4797-a1f9-a76b379da91c",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 404,
                      "message": "Not Found",
                      "requestId": "3793ab9a-5673-4c20-8580-c3240e3f3342",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 409,
                      "message": "Conflict",
                      "requestId": "a3d8f67a-a6d6-4a98-ad6c-ffa15711e862",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 500,
                      "message": "Internal Server Error",
                      "requestId": "a9b02586-d72b-41b3-be78-1d1f29e18312",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật điều kiện quyền",
        "tags": [
          "Attribute (Quản lý thuộc tính/điều kiện)"
        ]
      },
      "delete": {
        "operationId": "AttributeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID của điều kiện quyền (định dạng UUID)",
            "schema": {
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ResponseAttributeDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 400,
                      "message": "Bad Request",
                      "requestId": "a0b97693-ec90-4797-a1f9-a76b379da91c",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 404,
                      "message": "Not Found",
                      "requestId": "3793ab9a-5673-4c20-8580-c3240e3f3342",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 409,
                      "message": "Conflict",
                      "requestId": "a3d8f67a-a6d6-4a98-ad6c-ffa15711e862",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": {
                      "statusCode": 500,
                      "message": "Internal Server Error",
                      "requestId": "a9b02586-d72b-41b3-be78-1d1f29e18312",
                      "timestamp": "2026-05-01T07:02:21.515Z",
                      "path": "/attribute"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa điều kiện quyền",
        "tags": [
          "Attribute (Quản lý thuộc tính/điều kiện)"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS Pro API",
    "description": "Tài liệu API chi tiết cho dự án",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local environment"
    },
    {
      "url": "https://yourdomain.com",
      "description": "Staging"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Nhập mã token vào đây",
        "in": "header"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "object",
            "example": "attribute với name này đã tồn tại."
          },
          "requestId": {
            "type": "string",
            "example": "b4455dc9-313b-4914-8fe2-9e257abe29b8"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-30T11:46:33.862Z"
          },
          "path": {
            "type": "string",
            "example": "/attribute"
          }
        },
        "required": [
          "statusCode",
          "message",
          "requestId",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "ResponseAttributeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123"
          },
          "name": {
            "type": "string"
          },
          "_count": {
            "type": "object",
            "properties": {
              "conditions": {
                "type": "number"
              }
            },
            "required": [
              "conditions"
            ]
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CreateAttributeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "@description Tên định danh duy nhất của thuộc tính (slug).",
            "example": "department_code",
            "maxLength": 255
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateAttributeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "@description Tên định danh duy nhất của thuộc tính (slug).",
            "example": "department_code",
            "maxLength": 255
          }
        }
      }
    }
  }
}